리펙토링 진행중. readme 작성

+++ backend/src/main/java/com/ll/backend/domain/order/dto/OrderSummaryItemDto.java
@@ -1,11 +1,10 @@
 package com.ll.backend.domain.order.dto;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 
 public record OrderSummaryItemDto(
         String orderUid,
         int orderStatus,
         String orderStatusDisplay,
-        BigDecimal totalAmount,
+        long totalAmount,
         Instant orderedAt) {}
